requirements.txt
segno
pyzbar @ git+https://github.com/nmccann/pyzbar.git@master
Pillow
python-dotenv
dnspython
pynacl
opentimestamps-client
brotli
PyPDF2
qrcode[pil]
cryptography
